JAVADOCS completed

+++ src/main/java/com/michaelpirlis/inventorymanagement/InventoryController.java
@@ -26,18 +26,23 @@ import static com.michaelpirlis.inventorymanagement.models.Inventory.lookupPart;
 import static com.michaelpirlis.inventorymanagement.models.Inventory.lookupProduct;
 
 
+/** The Inventory Controller class launches the main GUI of the Inventory Management system.
+ * This contains the main tables and buttons for navigation, redirection, addition, modification,
+ deletion, and exiting the application. */
 public class InventoryController extends Application implements Initializable {
 
     @FXML private TableColumn<Object, Object> partIdColumn;
     @FXML private TableColumn<Object, Object> partNameColumn;
     @FXML private TableColumn<Object, Object> partInventoryColumn;
     @FXML private TableColumn<Object, Object> partPriceColumn;
+    @FXML private TableView<Part> allPartTable;
+
     @FXML private TableColumn<Object, Object> productIdColumn;
     @FXML private TableColumn<Object, Object> productNameColumn;
     @FXML private TableColumn<Object, Object> productInventoryColumn;
     @FXML private TableColumn<Object, Object> productPriceColumn;
     @FXML private TableView<Product> allProductTable;
-    @FXML private TableView<Part> allPartTable;
+
     @FXML private TextField partSearch;
     @FXML private TextField productSearch;
     public static Part modifyPart;
@@ -50,6 +55,10 @@ public class InventoryController extends Application implements Initializable {
         productTableSetup(allProductTable, productIdColumn, productNameColumn, productInventoryColumn, productPriceColumn);
     }
 
+    /** Method to configure the columns within the Part Table and assign values then called in the initializer.
+     * This method took an exceedingly long time to generate at Mark Kinkead made a correction
+     * in the video stating that the getters must be used. After hours, I realized the method worked
+     * in the video without the corrected steps. */
     static void partTableSetup(TableView<Part> allPartTable,
                                TableColumn<Object, Object> partIdColumn,
                                TableColumn<Object, Object> partNameColumn,
@@ -62,6 +71,9 @@ public class InventoryController extends Application implements Initializable {
         partPriceColumn.setCellValueFactory(new PropertyValueFactory<>("price"));
     }
 
+    /** Method to configure the columns within the Product Table and assign values then called in the initializer.
+     * After the issue with the partTableSetup, this was easy as it allowed duplicate code with changing to products
+     * instead of Parts.*/
     static void productTableSetup(TableView<Product> allProductTable,
                                   TableColumn<Object, Object> productIdColumn,
                                   TableColumn<Object, Object> productNameColumn,
@@ -84,6 +96,8 @@ public class InventoryController extends Application implements Initializable {
         stage.show();
     }
 
+    /** Generates a dialog box when no search results are found. Creating the specific look too multiple
+     * attempts. The dialog box by default appears dated and not optimized to display meaningful content.*/
     private void searchAlert() {
         Alert searchError = new Alert(Alert.AlertType.INFORMATION);
         searchError.setTitle("No Search Results Found");
@@ -93,6 +107,8 @@ public class InventoryController extends Application implements Initializable {
         searchError.showAndWait();
     }
 
+    /** Generates a dialog box when no selection is made in the table to modify.
+     * This was very straightforward after creating the searchAlert. Requiring a title and content text change. */
     private void noSelection() {
         Alert noSelection = new Alert(Alert.AlertType.INFORMATION);
         noSelection.setTitle("No Selection");
@@ -101,6 +117,9 @@ public class InventoryController extends Application implements Initializable {
         noSelection.showAndWait();
     }
 
+    /** Stores the search criteria in a string and checks for a matching name. If the search results are empty,
+     * then the process moves towards looking for a matching id. If nothing is found after both searches
+     * the alert appears. A try catch loop was added in case an unexpected value reaches the integer search. */
     @FXML
     private void searchParts() {
         String searchString = partSearch.getText();
@@ -127,6 +146,10 @@ public class InventoryController extends Application implements Initializable {
         partSearch.clear();
     }
 
+    /** Stores the search criteria in a string and checks for a matching name. If the search results are empty,
+     * then the process moves towards looking for a matching id. If nothing is found after both searches
+     * the alert appears. A try catch loop was added in case an unexpected value reaches the integer search.
+     * I created the part search first and this was a duplication changing parts to products. */
     @FXML
     private void searchProducts(){
         String searchString = productSearch.getText();
@@ -165,7 +188,7 @@ public class InventoryController extends Application implements Initializable {
 
     /** Checks for a selected Product from the table. If none is selected an alert appears,
      * otherwise it loads the modify product FXML file and form, sets up the scene, and displays the primary stage.
-     * Initially the screen progress regardless of being selected. I then realized I missed initializing modifyProduct
+     * Initially the screen progressed regardless of any selection. I then realized I missed initializing modifyProduct
      * before the null check. */
     @FXML
     private void modifyProductButton(ActionEvent event) throws IOException {
@@ -173,6 +196,7 @@ public class InventoryController extends Application implements Initializable {
 
         if (modifyProduct == null) {
             noSelection();
+
         } else {
                 Parent mainMenuParent = FXMLLoader.load(Objects.requireNonNull(getClass().getResource("modify-product.fxml")));
                 Scene mainMenuScene = new Scene(mainMenuParent);
@@ -193,7 +217,8 @@ public class InventoryController extends Application implements Initializable {
     }
 
     /** This method checks for a selected Part from the table. If none is selected an alert appears,
-     * otherwise it loads the modify product FXML file and form, sets up the scene, and displays the primary stage. */
+     * otherwise it loads the modify product FXML file and form, sets up the scene, and displays the primary stage.
+     * Similar to the modifyProductButton I needed to incorporate the null check.*/
     @FXML
     private void modifyPartButton(ActionEvent event) throws IOException {
         modifyPart = allPartTable.getSelectionModel().getSelectedItem();
@@ -209,7 +234,8 @@ public class InventoryController extends Application implements Initializable {
         }
     }
 
-    /** Created an alert to confirm part deletion. I expanded the scope to include the part name in the dialog prompt. */
+    /** Created an alert to confirm part deletion after the button is pressed.
+     * I expanded the scope to include the part name in the dialog prompt. */
     @FXML
     private void deletePartButton() {
         Part selectedPart = allPartTable.getSelectionModel().getSelectedItem();
@@ -226,22 +252,40 @@ public class InventoryController extends Application implements Initializable {
         }
     }
 
+    /** Created an alert to confirm product deletion after the button is pressed.
+     * I expanded the scope to include the product name in the dialog prompt. I also attempted to add
+     * the Associated Part to the modify button instead of the delete button and could not figure out why the prompt
+     * would not trigger with a size greater than 0. */
     @FXML
     private void deleteProductButton() {
         Product selectedProduct = allProductTable.getSelectionModel().getSelectedItem();
 
-        Alert confirmProductDeletion = new Alert(Alert.AlertType.CONFIRMATION);
-        confirmProductDeletion.setTitle("Confirm Product Deletion");
-        confirmProductDeletion.setHeaderText(null);
-        confirmProductDeletion.setContentText("Would you like to delete the product " + (selectedProduct.getName()) +" ?" +
-                "\nThis action is final and cannot be undone.");
+        if (selectedProduct.getAllAssociatedParts().size() != 0) {
+            Alert associatedPartDetected = new Alert(Alert.AlertType.INFORMATION);
+            associatedPartDetected.setTitle("Associated Part Detected");
+            associatedPartDetected.setHeaderText(null);
+            associatedPartDetected.setContentText("""
+                    This product contains associated parts.\s
 
-        Optional<ButtonType> result = confirmProductDeletion.showAndWait();
-        if (result.isPresent() && result.get() == ButtonType.OK) {
-            Inventory.deleteProduct(selectedProduct);
+                    Please use the modify button to remove parts,
+                    then save the changes to allow product deletion.""");
+            associatedPartDetected.showAndWait();
+
+        } else {
+            Alert confirmProductDeletion = new Alert(Alert.AlertType.CONFIRMATION);
+            confirmProductDeletion.setTitle("Confirm Product Deletion");
+            confirmProductDeletion.setHeaderText(null);
+            confirmProductDeletion.setContentText("Would you like to delete the product " + (selectedProduct.getName()) + " ?" +
+                    "\nThis action is final and cannot be undone.");
+
+            Optional<ButtonType> result = confirmProductDeletion.showAndWait();
+            if (result.isPresent() && result.get() == ButtonType.OK) {
+                Inventory.deleteProduct(selectedProduct);
+            }
         }
     }
 
+    /** On button press to Exit the application, the window closes.*/
     @FXML
     private void exitButtonClicked() {
         Platform.exit();
